home moved to views folder
